Moved KmzImage and related to its own namespace
